- moved deepCopyVisitor into containerFreeFunctions.hpp - fixed sparsityPattern.cpp include.
